The most significant changes involve the addition of localization to the application, which allows it to support multiple languages and cultures. This is achieved by injecting the `IStringLocalizer` interface into the controllers and using it to replace hard-coded strings with localized strings. The `IStringLocalizer` interface is injected into the `AnnouncementController`, `ChooseRoleController`, and `CollegeController` classes. The localized strings are then used in place of hard-coded strings in these classes. The `TempData["Message"]` and `TempData["Error"]` values in the controllers are also replaced with localized strings.
The changes also include the addition of the `Microsoft.Extensions.Localization` namespace, which provides the `IStringLocalizer` interface. The constructor of each controller is updated to accept an instance of `IStringLocalizer`. This instance is then assigned to a private field, which is used to fetch the localized strings.

The `AnnouncementController.ar.resx` file, which contains the localized strings for the Arabic language, is updated with the new strings. This file is used by the `IStringLocalizer` interface to fetch the localized strings. The `resheader` values in the `AnnouncementController.ar.resx` file are updated to version 2.0. This indicates that the file is using the version 2.0 of the ResX schema.

The `AnnouncementController.en.resx` file was updated from version 1.3 to version 2.0. The new version includes additional comments and examples, as well as new data entries for "announcNotFound", "collegeNotFound", "deptNotFound", "errorLoadCollege", and "errorLoadDept". The reader and writer versions were also updated to 4.0.0.0.

The `BaseController.ar.resx`, `BaseController.en.resx`, and `BoardsController.ar.resx` files were all removed. These files previously contained version 1.3 of the Microsoft ResX Schema. The `BoardsController.en.resx` file was completely removed.

The `ChooseRoleController.ar.resx` and `ChooseRoleController.en.resx` files were updated. The version of the Microsoft ResX Schema was changed from 1.3 to 2.0. The XML schema was also updated to reflect this change. Additionally, new data entries were added to these files. These entries include "multiRoles", "noPermission", "selectRole", and "unableLoadUser", each with their respective values.

The changes in both `CollegeController.ar.resx` and `CollegeController.en.resx` files are related to the upgrade of the ResX Schema from version 1.3 to version 2.0. This is indicated by the addition of the line `<resheader name="version">2.0</resheader>` and the removal of the line `<resheader name="version">1.3</resheader>`. The changes also include the addition of new data elements in both files. These new data elements are related to the messages displayed to the user in different scenarios. For example, the message displayed when a college cannot be deleted because it has existing departments, when a college is deleted successfully, when a college is not found, and when the user is asked to delete the departments first. The changes also include the addition of new XML schema definitions for the new data elements. These definitions specify the structure of the new data elements, including their names, types, and other attributes. The changes also include the removal of some old XML schema definitions. These definitions were related to the old version of the ResX Schema and are no longer needed in the new version.